added index.js file to routes directory. Moved code for api GET, POST, and DELETE routes to notes.js and added dependencies, exports, and app
