+++ routes/index.js
@@ -1,0 +1,10 @@
+const express = require('express');
+
+const notesRouter = require('./notes');
+
+const app = express();
+
+app.use('/notes', notesRouter);
+
+module.exports = app;
+
